Move current layout to API, add new default.
New default layout is for pages without the API sidebar. Moved the
existing layout to be the api pages layout. Update existing MD pages
accordingly.

Rip out some of the custom CSS that was added to bootstrap.css into the
customization file.

Add in the responsive CSS.

Add HTML5 tags for header, footer, and the default layout.