Update "Buy" Links in Developer site
Most of the links will just redirect to main marketing site. Most of
these were redundant amazon buttons that are in both places.

The `-m` page was a mobile only page that predated the mobile optimized
marketing site. That page now redirects to the canonical buy page for
the dev kit.

The SD card page now points directly to the order page on
account.radiusnetworks.com.